refactor component names
