+++ src/app/(main)/order/page.tsx
@@ -1,5 +1,5 @@
 import { Providers } from '@/app/providers';
-import CreateOrderForm from '@/components/orders/CreateOrderForm';
+import CreateOrderForm from '@/components/orders/create-order-form';
 import { getAvailableCommodities } from '@/server-actions/commodities';
 
 const Page = async () => {
