+++ src/server-actions/orders.ts
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from 'next/cache';
 
-import type { OrderFormSchema } from '@/components/orders/OrderForm';
+import type { OrderFormSchema } from '@/components/orders/order-form';
 import { createOrder } from '@/server/orders';
 
 export const createOrderServerAction = async (order: OrderFormSchema) => {
